Remove ghcVers from safe-exceptions
The build matrix was really just a proxy for testing that we have
working args for each ghcVers. But we now do that as part of the
default CI. We really don't need to test that safe-exceptions,
or any other 'default' lib, works for multiple ghcs. We should only
test that if we want to explicitly ensure a specific lib builds
with specific ghc vers (e.g. if liquidhaskell builds for multiple
ghcs).