Add validate back to ghc alias
Turns out we want this, and it's working, so should have it back in.
This seems to be a more robust way to run validation i.e. I have
had trouble running ./validate due to a warning regarding
having FORTIFY_SOURCE=2 and -O0 simultaneously (the former triggers
-Wcpp due to it requiring optimizations, and -Werror makes this a fatal
error.

Efforts to disable FORTIFY / turn on optimizations in ./validate have
failed, so we stick with hadrian for now.