Fix ghc 9.6.2/3 hls
Also swap CI's --dry-run for a build with hls. It turns out CI
__does__ have enough memory to load hls when it is the only tool.
This is important as ghc 9.6.2 and 9.6.3 had broken hls's.

This occurred as we incorrectly assumed --dry-run would report
such errors (it did not), so actually building it appears to be
the best way forward, assuming we do not run out of storage.