Explicitly note ghc981 test shell not loading tools
The ghc981 test (ci) shell was not loading all tools. This was probably
unintentional, as all args should be passed through be default, with
overrides where noted.

That said, it turns out all of the other tools have poor caching, so
we explicitly disable them.

We also remove ghc981 from HLS CI, which is sad because we actually
want this (liquidhaskell shell), but caching is poor, so oh well.

Finally, fix a minor error where general test shell was not loading
ormolu. This is minor since fourmolu was loading it anyway, but it
should be explicitly enabled.