Update GHC 9.10.1 to get new Cabal 3.12.
This updates the GHC 9.10.1 hash to get the new Cabal. This is needed
to build the cabal repo. Note that CI seems to have a separate
problem that I have been unable to reproduce locally.

Nevertheless we merge this, as it is a clear upgrade, even if CI
is still not totally fixed.